DCAC-80: Reduce test execution time.
